Added Android gesture (Not working well yet) because of Johanna.
