+++ GoKitchen/res/layout/activity_gesture.xml
@@ -1,5 +1,6 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.gesture.GestureOverlayView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/gestures"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/activity_vertical_margin"
@@ -7,6 +8,12 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="com.app.gokitchen.GestureActivity" >
+    
+    <TextView
+        android:id="@+id/textView1"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="TextView" />
 
     <View
         android:id="@+id/viewGesture"
@@ -15,4 +22,4 @@
         android:layout_centerHorizontal="true"
         android:layout_centerVertical="true" />
 
-</RelativeLayout>
+</android.gesture.GestureOverlayView>
