+++ GoKitchen/src/com/app/gokitchen/util/DrawView.java
@@ -1,0 +1,171 @@
+package com.app.gokitchen.util;
+
+import com.app.gokitchen.GestureActivity;
+import com.app.gokitchen.OCRActivity;
+
+import android.content.Context;
+import android.content.Intent;
+import android.graphics.Bitmap;
+import android.graphics.Canvas;
+import android.graphics.Color;
+import android.graphics.Paint;
+import android.graphics.Path;
+import android.util.Log;
+import android.view.MotionEvent;
+import android.view.View;
+import android.view.GestureDetector.OnDoubleTapListener;
+import android.view.GestureDetector.OnGestureListener;
+
+public class DrawView extends View { //implements OnGestureListener, OnDoubleTapListener{
+
+   public int width;
+    public  int height;
+    private Bitmap  mBitmap;
+    private Canvas  mCanvas;
+    private Path    mPath;
+    private Paint   mBitmapPaint;
+    Context context;
+    private Paint circlePaint;
+    private Path circlePath;
+
+    public DrawView(Context c) {
+        super(c);
+        context=c;
+        mPath = new Path();
+        mBitmapPaint = new Paint(Paint.DITHER_FLAG);
+        circlePaint = new Paint();
+        circlePath = new Path();
+        circlePaint.setAntiAlias(true);
+        circlePaint.setColor(Color.BLUE);
+        circlePaint.setStyle(Paint.Style.STROKE);
+        circlePaint.setStrokeJoin(Paint.Join.MITER);
+        circlePaint.setStrokeWidth(4f);
+    }
+/*
+    @Override
+    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
+        super.onSizeChanged(w, h, oldw, oldh);
+
+        mBitmap = Bitmap.createBitmap(w, h, Bitmap.Config.ARGB_8888);
+        mCanvas = new Canvas(mBitmap);
+    }
+
+    @Override
+    protected void onDraw(Canvas canvas) {
+        super.onDraw(canvas);
+
+        canvas.drawBitmap( mBitmap, 0, 0, mBitmapPaint);
+        canvas.drawPath( mPath,  mPaint);
+        canvas.drawPath( circlePath,  circlePaint);
+    }
+
+    private float mX, mY;
+    private static final float TOUCH_TOLERANCE = 4;
+
+    private void touch_start(float x, float y) {
+        mPath.reset();
+        mPath.moveTo(x, y);
+        mX = x;
+        mY = y;
+    }
+
+    private void touch_move(float x, float y) {
+        float dx = Math.abs(x - mX);
+        float dy = Math.abs(y - mY);
+        if (dx >= TOUCH_TOLERANCE || dy >= TOUCH_TOLERANCE) {
+            mPath.quadTo(mX, mY, (x + mX)/2, (y + mY)/2);
+            mX = x;
+            mY = y;
+
+            circlePath.reset();
+            circlePath.addCircle(mX, mY, 30, Path.Direction.CW);
+        }
+    }
+
+    private void touch_up() {
+        mPath.lineTo(mX, mY);
+        circlePath.reset();
+        // commit the path to our offscreen
+        mCanvas.drawPath(mPath,  mPaint);
+        // kill this so we don't double draw
+        mPath.reset();
+    }
+
+    @Override
+    public boolean onTouchEvent(MotionEvent event) {
+        float x = event.getX();
+        float y = event.getY();
+
+        switch (event.getAction()) {
+            case MotionEvent.ACTION_DOWN:
+                touch_start(x, y);
+                invalidate();
+                break;
+            case MotionEvent.ACTION_MOVE:
+                touch_move(x, y);
+                invalidate();
+                break;
+            case MotionEvent.ACTION_UP:
+                touch_up();
+                invalidate();
+                break;
+        }
+        return true;
+    }
+
+	@Override
+	public boolean onSingleTapConfirmed(MotionEvent e) {
+        Log.d(DEBUG_TAG, "onSingleTapConfirmed: " + e.toString());
+		return true;
+	}
+
+	@Override
+	public boolean onDoubleTap(MotionEvent e) {
+		Log.d(DEBUG_TAG, "onDoubleTap: " + e.toString());
+		return true;
+	}
+
+	@Override
+	public boolean onDoubleTapEvent(MotionEvent e) {
+		Log.d(DEBUG_TAG, "onDoubleTapEvent: " + e.toString());
+		return false;
+	}
+
+	@Override
+	public boolean onDown(MotionEvent e) {
+		Log.d(DEBUG_TAG,"onDown: " + e.toString());
+		try{
+			Intent intent = new Intent(GestureActivity.this, OCRActivity.class);
+			startActivity(intent);
+		}catch (Exception e1){}
+		return true;
+	}
+
+	@Override
+	public void onShowPress(MotionEvent e) {
+		Log.d(DEBUG_TAG, "onShowPress: " + e.toString());		
+	}
+
+	@Override
+	public boolean onSingleTapUp(MotionEvent e) {
+		Log.d(DEBUG_TAG, "onSingleTapUp: " + e.toString());
+		return true;
+	}
+
+	@Override
+	public boolean onScroll(MotionEvent e1, MotionEvent e2, float distanceX, float distanceY) {
+		Log.d(DEBUG_TAG, "onScroll: " + e1.toString()+e2.toString());
+		return true;
+	}
+
+	@Override
+	public void onLongPress(MotionEvent e) {
+		Log.d(DEBUG_TAG, "onLongPress: " + e.toString());		
+	}
+
+	@Override
+	public boolean onFling(MotionEvent e1, MotionEvent e2, float velocityX, float velocityY) {
+        Log.d(DEBUG_TAG, "onFling: " + e1.toString()+e2.toString());
+        return true;
+	}*/
+}
